Splitted the callback and the processor of FedEMA
